+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/kafka/AbstractAzureKafkaOAuth2AutoConfigurationTests.java
@@ -4,6 +4,7 @@ package com.azure.spring.cloud.autoconfigure.kafka;
 
 import java.util.Map;
 
+import com.azure.identity.providers.jdbc.implementation.enums.AuthProperty;
 import com.azure.spring.cloud.autoconfigure.context.AzureGlobalPropertiesAutoConfiguration;
 import com.azure.spring.cloud.autoconfigure.context.AzureTokenCredentialAutoConfiguration;
 import org.junit.jupiter.api.Test;
@@ -33,8 +34,8 @@ abstract class AbstractAzureKafkaOAuth2AutoConfigurationTests {
     protected static final String SPRING_BOOT_KAFKA_PRODUCER_PROPERTIES_PREFIX = "spring.kafka.producer.properties.";
     protected static final String SPRING_CLOUD_STREAM_KAFKA_PROPERTIES_PREFIX = "spring.cloud.stream.kafka.binder.configuration.";
     protected static final String SPRING_CLOUD_STREAM_KAFKA_CONSUMER_PROPERTIES_PREFIX = "spring.cloud.stream.kafka.binder.consumer-properties.";
-    protected static final String CLIENT_ID = "azure.credential.client-id";
-    protected static final String MANAGED_IDENTITY_ENABLED = "azure.credential.managed-identity-enabled";
+    protected static final String CLIENT_ID = AuthProperty.CLIENT_ID.getPropertyKey();
+    protected static final String MANAGED_IDENTITY_ENABLED = AuthProperty.MANAGED_IDENTITY_ENABLED.getPropertyKey();
 
     protected final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
             .withConfiguration(AutoConfigurations.of(AzureEventHubsKafkaOAuth2AutoConfiguration.class,